use size_t

+++ crawlserv/src/Data/Data.hpp
@@ -58,7 +58,7 @@ namespace crawlservpp::Data {
 			int _i;
 			unsigned int _ui;
 			long _l;
-			unsigned long _ul;
+			size_t _ul;
 			double _d;
 		};
 
@@ -82,7 +82,7 @@ namespace crawlservpp::Data {
 		Value(int value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_i = value; } // int
 		Value(unsigned int value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_ui = value; } // unsigned int
 		Value(long value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_l = value; } // long
-		Value(unsigned long value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_ul = value; } // unsigned long
+		Value(size_t value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_ul = value; } // size_t
 		Value(double value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_d = value; } // double
 		Value(const std::string& value) { this->_isnull = false; this->_overflow = _if_too_large::_error; this->_s = value; } // string
 		void clear() { Value(); }
